feat: implementqcao de testes unitarios para o dominio clinic

+++ Wpm.SharedKerbel/ValueObjects/Weight.cs
@@ -13,7 +13,7 @@
         //Ele permite que um valor do tipo double seja automaticamente convertido em um objeto Weight, sem precisar chamar o construtor explicitamente.
         //O método estático implicit operator Weight(double value) é um atalho do compilador que transforma double automaticamente em Weight,
         //permitindo usar valores numéricos diretamente no domínio, mas garantindo que eles sempre passem pelas regras de validação do Weight.
-        public static implicit operator Weight(double value)
+        public static implicit operator Weight(decimal value)
         {
             return new Weight(value);
         }
